+++ lib/eve_online.rb
@@ -26,8 +26,6 @@ require 'eve_online/xml/character_wallet_journal'
 
 require 'eve_online/xml/corporation_market_orders'
 
-require 'eve_online/xml/models/account_type_object'
-
 # ESI API
 require 'eve_online/esi/base'
 
@@ -51,9 +49,14 @@ require 'eve_online/esi/corporation_industry_jobs'
 
 require 'eve_online/esi/alliance'
 require 'eve_online/esi/alliance_icon'
+
+require 'eve_online/esi/dogma_attributes'
+require 'eve_online/esi/dogma_attribute'
+
 require 'eve_online/esi/server_status'
 require 'eve_online/esi/races'
 require 'eve_online/esi/bloodlines'
+require 'eve_online/esi/factions'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -76,8 +79,11 @@ require 'eve_online/esi/models/corporation'
 require 'eve_online/esi/models/alliance'
 require 'eve_online/esi/models/alliance_icon'
 
+require 'eve_online/esi/models/dogma_attribute'
+
 require 'eve_online/esi/models/race'
 require 'eve_online/esi/models/bloodline'
+require 'eve_online/esi/models/faction'
 
 # SDE
 require 'eve_online/sde/base'